Clean it up to just the methods
